AI/requirements.txt

flask
opencv-python
numpy
scikit-learn
gunicorn
torch
torchvision
sentence-transformers
Pillow

